jdmarker.c: stop at 61th scan
According to report accompanying issue LJT­01­003, it's possible to generate a
small .jpg file that'll take eons to process and document that describes it
says it's possible to limit that by stopping processing at particular point.

This patch builts that into jdmarker.c, so client application doesn't need to
worry about it.

Signed-off-by: Piotr Dałek <git@predictor.org.pl>